requirements.txt
pdfminer.six
rich==10.2.0
